Updated openstack/openstack Project: openstack/cinder  91f18814405990d799c7a6bd37deef9254f7cdd1
SMBFS: Fix retrieving total allocated size

Currently, the Windows SMBFS driver parses all the VHD/X images
present on the configured share in order to retrieve the allocated
size. Not only that this can be slow when having many volumes, but
it can also fail in certain environments when the images are in use.

The Linux SMBFS driver uses 'du', which gives incorrect values in
case of VHD/X images.

In order to avoid this, the driver can keep track of the allocated
size according to each share, the most basic solution being a JSON
stored in a file.

As the Windows SMBFS driver inherits the Linux SMBFS driver, only
overwriting the os specific methods, the fix is applied to the base
SMBFS driver.

DocImpact

Closes-Bug: #1425100

Change-Id: I14aa7b001605ae14fe0b6d0a45ee6b1abf8c4f60