Updated openstack/openstack Project: openstack/swift  f26af37bc2735ba9b6215cbe20a9517cefbed9c5
Read the response body, if response status is greater than 300.

internal_client was not reading response if response status is not 200.
So proxy server treats this as client disconnect and logs 499 in log file.
This patch fixes that by reading response if response status is greater
than or equal to 300 and in acceptable statuses.

Closes-Bug: #1364752

Change-Id: I0512a25895da583956f76031e3c5de5c970bce01