Updated openstack/openstack Project: openstack/cinder  568a4c7bd0cd8870abbac18776eacef3e918e000
Brocade driver add_zone optimization

Check whether the new zone we want to add/create
has the same zone name and members with the existing
zone on the switch. If they are same, no need to
delete and re-add it. This is significant if we
create volume from image or upload volume to image
in batch.

Implements: blueprint brocade-add-zone-optimization
Change-Id: I257ec25fed9fab4b3a85a0f200c2ee426c8e6f8c