Updated openstack/openstack Project: openstack/swift  f4e1f9fbf162f894653a687400e6c5f459bf4167
On py3, use dnspython3 dependency, not dnspython

Update requirements to install dnspython3 on Python 3 and dnspython on
Python 2. dnspython is not compatible with Python 3, and dnspython3
is not compatible with Python 2.

Add a requirement on pbr to ensure that pbr 1.6 or newer is
installed. pbr 1.0 or newer is required to support environment
markers in requirements which are now used to select dnspython or
dnspython3 depending on the Python major version.

Note: pbr 1.0 is enough, but OpenStack global requirements asks to
use "pbr>=1.6".

Change-Id: Ie5d437cd396972d4143ffffb96ec3c289e221b08