Updated openstack/openstack Project: openstack/barbican  7ff4ee1c1075968ed4579568cec348e8a32c4c7a
Allow null values for secret names

This secret allows null values for names in secrets as discussed a
couple of cycles back.

Since the code for the mocked client in the functional tests actually
filters out null values, it is currently not possible to write a
functional test for this in this functional test code base. However, it
will be written as part of the python-barbicanclient functional test
suite.

Change-Id: I68330d1d1b13e1bbb6436ca4f6684a76af3babd3
Related-bug: #1425667
Related-bug: #1376469