Updated openstack/openstack Project: openstack/barbican  e2f4278f93816da89b5a188c1b1f9cffbba99502
Allow null content type for orders

For both symmetric [1] and assymetric [2] orders the content type
is defaulted to 'application/octet-stream'. On the other hand, for
certificates the value is completely ignored. Based on the previous two
facts, we can pass null values for the content type via the API, as we
had discussed a couple of cycles ago. Thus expecting Barbican's defaults
to act appropriately.

[1] https://github.com/openstack/barbican/blob/master/barbican/tasks/resources.py#L285
[2] https://github.com/openstack/barbican/blob/master/barbican/tasks/resources.py#L295

Change-Id: I98a558862eed039ed6f33fbee474bdb86767313e
Related-bug: #1425667