Updated openstack/openstack Project: openstack/nova  0da0971cc44b93110032a3b382614f3f84297951
Deprecate local conductor mode

Way back when we started no-db-compute, we introduced the conductor service
and provided a "local mode" which made conductor's API layer a passthrough to
the manager and thus the database. This eliminated the need to move to
conductor for deployments that weren't ready or had other blockers.

Now, conductor is a pretty critical part of nova's infrastructure and
continuing to allow both modes invites issues (such as bug 1505471). This
patch starts the timer on deprecation for this mode of operation so that
we can remove it in the future.

UpgradeImpact: A deployment mode is being deprecated and will require any
existing operators using this to deploy conductor before removal.

Change-Id: Ifbd5578dcc53b2117674db2016e5d882a82866aa