Updated openstack/openstack Project: openstack/heat-translator  33cf3cc1ed0800827bffec715098d746dd042e8a
Refactor dynamic types mapping feature to use importlib

It was discovered that the Python imp.find_module() method behaves
differently on Mac environments. Sometimes it does not find the
specified module even if it does exist on the sys.path.

To fix this for all platforms, the code was refactored to use the
importlib.import_module() method. This module looks to correctly
work on both platforms, Linux and Mac OSX.

Change-Id: I521cd430dc04f60e44c192cb3194a0ae07efacb7