Updated openstack/openstack Project: openstack/cinder  ea222055cef85f3ecc98b9ca509ad3291f985a6d
Add protocol to help of glance_api_servers option

Help for glance_api_servers option lacked information that it is
possible to specify protocol there (http or https) and suggested that
it isn't possible to use HTTPS in connection from Cinder to Glance.
This commit expands the message to include such information.

Change-Id: I5afa33193b6dda2477e5a28fa042e625d449257b
Closes-Bug: 1509065