Updated openstack/openstack Project: openstack/python-openstackclient  1cf09be91366224dfd95ab7013151e0510012dbb
Use fake server name instead of id when testing "server_name" param.

"server create" command takes a server name parameter. The server id is
generated by openstack.

When we intended to pass a server name to the unit tests of "server create"
command, we passed server id. It won't be any problem because the fake
server id is also a string, but we should pass a fake server name because
we have a string to fake one.

Change-Id: I9944f0ea2a6c457e4fad8215a54778bca08965ab