Updated openstack/openstack Project: openstack/keystone  434cc5bea32350e1dd423d4393f8c401812fe871
Merge tag '8.0.0'

This is a null-merge of the 8.0.0 release tag back into the master
branch so that the 8.0.0 tag will appear in the git commit history of
the master branch. It contains no actual changes to the master branch,
regardless of how our code review system's UI represents it. Please
ask in #openstack-infra if you have any questions, and otherwise try
to merge this as quickly as possible to avoid later conflicts on the
master branch.

Change-Id: I7fb57bcb6ea824cd62e08ba0d4e728ce0ea773e1