Updated openstack/openstack Project: openstack/nova  e8004142c9b86c74bcd36f30944b5ce3b1d0a886
Remove flavors.get_all_flavors() only used in tests

This method is only used in tests, sometimes even in orphaned code.
This method really just calls the object method FlavorList.get_all()
and wraps the result in a dict.

Lets remove it an replace the few remaining places where it actually
was used with direct calls to the object method.

Change-Id: I990f81af3d24a5b98d1905d44262c138fd2a0283