Updated openstack/openstack Project: openstack/nova  b12e72ff230fd455d98cd3d4cc33fceee40f681a
Fix missing obj_make_compatible() for ImageMetaProps object

This object was never added with an obj_make_compatible() handler for
the many versions we've had since its introduction. This adds that,
rules for the existing versions, and a test for the future.

Note that version 1.6 added values to an enum, so this validates that
an exception is raised during backporting if that field has a value
that would be invalid for the requested version.

Change-Id: I64d40b6fe5e738cf32537c73b9fec57a7fbcdf2d