Updated openstack/openstack Project: openstack/governance  0a4d1902df9d9a1b67be2878ece00c8481b63da3
Add kolla-mesos deliverable repo to Kolla

The kolla-mesos repository was majority voted accepted by the core
reviewer team of Kolla to be adopted into the Kolla governance.
The purpose of this repository is to orchestrate a Kolla deployment
in addition to Ansible but instead with Mesos.  If this experiment
fails, the repository will be filed in the attic.

The core reviewer team has committed to reviewing the changes for
this repository and learning how mesos operates.  Angus Salkeld
of Mirantis has committed 4 full time engineers to work on this
project.

The structure of the governance will be the addition of a team
kolla-mesos-core with only 2 people from Mirantis and the existing
core team from Kolla.  This will improve information transfer and
training of the existing core team for Mesos.  The short term goal
will be to retire this group and move the folks into the main
kolla-core team once they have met our informal requirements for
review output and quality, IRC participation, code generation, etc.

To avoid any code dump, this will be a fresh cookiecutter repository
that Angus will merge with project-config separately.

Change-Id: I902d4c31c13d4941a470553e8013ac7692462f1a