Updated openstack/openstack Project: openstack-infra/project-config  166c58ff9556e316593d734d1288c3a29f7b47e3
Add dsvm into job names for Kolla

clarkb suggested dsvm be in the job names for Kolla to follow the
job naming convention. In the case of the experimental job we have
recently merged, that it as a requirement for the name of the job.

To keep consistency within the project and within the OpenStack
community we are renaming all of the gates that use the dsvm's to the
convention that all other jobnames follow.

Since the tox paramaters do not change in the jobs/kolla.yaml, this
requires no change on the Kolla side.

Change-Id: I9a22434d99b4deb7936ecaf7b714d0dc41666367