Updated openstack/openstack Project: openstack/heat  0e47c75c95c31d93afec2be841da45ef11c77dcc
Do not resolve keystone values if it is None

In patch Ic7a3001ab271cad12f4fe45c0930562070a0d600 were added new
functions for resolving keystone values. However it may lead to an error,
when the user does not specify the related property and it's None.
These functions can not resolve None value and fail.
This patch adds check for None values.

Change-Id: I667e61e9553a92703dbc5a37a311d93c4a377d4b
Closes-Bug: #1509354