Updated openstack/openstack Project: openstack/horizon  8b4fc07267ab7b414a963d465e93fc6e905039f7
Add a Delete Modal Service for deleting entities

This provides a reusable modal service for deleting
entities. It can be used for deleting single or multiple
objects.

It requires that the backing API Service allow for
suppressing of errors so that success and error messages
are shown only once when working with multiple objects.

It's used in the patch
https://review.openstack.org/#/c/234873/

Change-Id: I9d6410c0d4a692db4044661254a9914eef79e3ed
Partially-Implements: blueprint ng-flavors