Updated openstack/openstack Project: openstack/horizon  f8ee84ba41a3dc56b231db6ef81e3cd7b569d5c1
Add STATUS_DISPLAY_CHOICE which matches Nova vm state

While working on another defect I noticed the value Horizon is using
in the status tuple for an instance which was soft deleted did not
match the string Nova is using for vm state.
Horizon is currently using the value "soft_deleted" in the
STATUS_DISPLAY_CHOICES tuple.  Nova is using the string
"soft-delete".  A more consistent approach in Nova would be to use
the string "soft_deleted" however changing this in Nova may
inadvertently change a REST API.
Adding a choice for "soft-delete" to Horizon enables Horizon to
translate the vm state correctly now.  Removing the existing choice
of "soft_deleted".

Change-Id: Icefd458bc3633c6fc8103e39966b9916722ff97f
Closes-Bug: #1506177