Updated openstack/openstack Project: openstack-infra/shade  b7c96d3a5a39a76f370a5ef0201b1a823fe1bcde
Let os-client-config handle session creation

More things than shade need consistent Session creation, so we moved
a copy of the shade code into os-client-config to allow python-*client
to consume the same logic. Now that it's there, consume it in shade and
remove the need for shade to know/duplicate the logic.

Depends-On: Ide4c613cc143a0b8a3f36130989b57e808b2530f
Change-Id: I86cdb3cdd2710ef302520184ccfcb1605384f706