Updated openstack/openstack Project: openstack/oslo.config  1db229424eaeb7e5e21d50bdb8ec708a22a97618
Remove "Kept for backward-compatibility" comment

There is not plan to deprecate, or even remove, the following specialized classes:

* StrOpt
* IntOpt
* FloatOpt
* ListOpt
* DictOpt
* MultiStrOpt

So remove the comment "Kept for backward-compatibility with options
not using Opt directly" to avoid any confusion.

These classes are widely used in OpenStack, whereas the cfg.Opt class
is not used directly.

Change-Id: I887dea6bcd1a6c17a3a40061a5c3c78104ed0046