Updated openstack/openstack Project: openstack/neutron-fwaas  18793f62c38bfe58184ba47e65c8ed5a3f1e8da4
Set ZUUL_BRANCH using BRANCH if not available

In most cases, ZUUL_BRANCH is set in gate, but not all the time. One
exception is e.g. periodic stable jobs, where there is no actual patch
to validate, hence no ZUUL_REF or ZUUL_BRANCH is set. In that case, we
need to help zuul-cloner to determine the proper branch to fetch
neutron.

It could be achieved by e.g. hardcoding the branch using --zuul_branch
CLI argument passed into zuul-cloner. In this patch, I instead set
ZUUL_BRANCH to BRANCH that seems to be always present in all kinds of
jobs. It is then used by zuul-cloner to choose the correct branch.

This will be needed in the future when we spin out stable/mitaka and run
stable periodic jobs for neutron-fwaas.

Change-Id: I360239a1443f8d4e16210a1ef743912e1bf85f41