Updated openstack/openstack Project: openstack/trove-integration  bce62b3da9c69a7a4f6ea33116261f853e1c9c26
Fix redstack (devstack changes broke ceilometer)

A recent commit in devstack (https://review.openstack.org/#/c/219207 )
now validates that all post-processed conf files specified in local.conf
exist.

We need to conditionally add the ceilometer.conf to this file, since it
may or may not be present.

Enabling Ceilometer was also broken, as it's now a member of the 'big
tent' so this was fixed as well

Two additional options were added to enable ceilometer notifications in
Nova and Cinder.  These can be activated by setting the following on
the command line or in options.rc (or ~/.redstack.options.rc).

CEILOMETER_NOVA_CONF=True
CEILOMETER_CINDER_CONF=True

Change-Id: I4181d751a1c35b581c7d56e157324d04c4ddec6b
Closes-Bug: #1513515