Updated openstack/openstack Project: openstack/governance  dce05e06de26d2712ef23b534f76fd228d8dd028
Introduce assert:supports-upgrade tag

The assert:supports-upgrade tag defines a standard minimum for
upgrade capabilities in OpenStack services. Project teams may
assert that they follow at least this policy.

This is an essential bit of information to communicate to our users,
and a key factor in assessing a given project maturity.

Change-Id: Iad564e22cd008a46f22a55204ee6780ef9c9b9dd