Updated openstack/openstack Project: openstack/governance  bbd929ecfc3ab10bb14e1f7d77eeb79a3f1b19c4
Introduce assert:supports-rolling-upgrade tag

The assert:supports-rolling-upgrade tag defines a standard minimum for
what we consider to be "rolling upgrade" capabilities in OpenStack
services. Project teams may assert that they follow at least
this policy.

This is an essential bit of information to communicate to our users,
and a key factor in assessing a given project's maturity.

Change-Id: I7ba9cdbe1b6ba026da374dd66e30611133538e91