Updated openstack/openstack Project: openstack/governance  c4f6902287df62f4f53416e12cb47f9af0184110
Rename Ceilometer projects to Telemetry

The Ceilometer team and project took a lot of new deliverable recently under
its umbrella, such as Gnocchi and Aodh, and the Ceilometer term does not
identify the team and project correctly anymore.

During discussion at the Mitaka summit, it appeared that emphasizing the name
Telemetry as the name of the group of people and deliverable they produce is
way more accurate than just "Ceilometer", which is now only a deliverable
getting smaller and smaller.

Change-Id: I99d6c2416b8d395935daec945cfd4ae686ce76e9