Updated openstack/openstack Project: openstack/keystonemiddleware  508cf14355f432aef7b182c60ad68787f0f4c1e2
Separate setting catalog on headers from others

A minor cleanup to make setting catalog headers distinct from the
regular headers. This makes it easier to see the difference between the
headers that are set for the service token and the user token as we only
set service catalog for user token.

Change-Id: I696e826a9983965258cf4cf518ab399723d27649