Updated openstack/openstack Project: openstack-infra/subunit2sql  028db33979cd4ab704eb81da539c028dccc5db3e
Add option to specify global test attr prefix in subunit2sql

This commit adds a CLI/config option to specify a prefix to identify
global test attrs. If specified it will treat the test attrs pulled
from the subunit stream as a comma seperated list and add a row, if
one doesn't exist, to the test_metadata table for each attr that
starts with the prefix provided. An example use case for this is the
idempotent id attr that tempest uses to track tests across renames
and refactors.

Change-Id: Ice14d42a53efa41fb0eb06bd4d1fe21c5a5a9a01