Updated openstack/openstack Project: openstack-infra/subunit2sql  588bfdeb0596c71b9b74799dcead2993ba9f7824
Enable python 3 testing support

In an effort to being really supporting using subunit2sql in a python3
environment this patch fixes all the issues with python3
incompatibility that show up when running tests. However, because the
test coverage is limited for certain parts there are still probably
pieces which don't work with python 3.

Change-Id: I160d1675069836fa514547c31866b22bcaf29dcb