Updated openstack/openstack Project: openstack/tempest  c8f2fda7851e3e7a1bdb96850fa56d427baea93d
Apply a naming rule of GET to v2 keystone clients

[GET /resources] methods should be "list_<resource name>s"
or "show_<resource name>", so this patch applies the rule
to v2 keystone clients.
This patch changes some v3 parts also because some tests are
shared between v2 and v3 and the method names should be the same.

Partially implements blueprint consistent-service-method-names

Change-Id: Iea7dcb8839d0dfc300295f3e534d54224f9ec31c