Updated openstack/openstack Project: openstack/neutron  87ce4868caea9b7f1dd79f575004ee152ffa7eb0
Add new config option for IPv6 Prefix Delegation

With the impending deprecation of the default subnetpool configuration
options, IPv6 PD needs to be enabled with a new config option.

This patch adds the 'ipv6_pd_enabled' option to neutron.conf, and makes
all of the necessary changes for its use.

DocImpact

Change-Id: I43486c5a13ee2ff0097355afe7e1f3ef8794b185
Closes-Bug: 1501835