Updated openstack/openstack Project: openstack-infra/storyboard-webclient  6e6d12160284369ab4cb6606bfd772c6a0068225
Move project group admin and streamline list page

The options to add, edit and delete project groups used to be hidden away on
the admin submenu. This was unintuitive because the other pages (stories and
projects) did not work this way. Moreover, the admin detail page had no simple
option to view a project group's projects and stories at the same time,
while the non-admin list page did not show a group's projects. This patch
simplifies things: there is now one Project Groups list page, which links to a
detail page, both of which have edit options for admin.

Change-Id: Id9ccb22b54158b03a50fb50fb609880e0c58c1d6