Updated openstack/openstack Project: openstack-infra/project-config  c34607d6281df6486ddd5acbf6c90693aadb10d2
layout: Fix regexp for puppet-openstack integration jobs

The change introduced in 295320dc78b20e44c23209766c17bf0677de46d0
update the regexp of puppet-openstack integration job, but matches
a non-existing job (scenario00(1|2)--dsvm). The problem behind that
is now all puppet-openstack integration jobs are voting... and we don't
want that for the moment.

Change-Id: Iceaafefc6202f8182526e8b4461a556cc06d918a