Updated openstack/openstack Project: openstack/python-novaclient  903538ee59dbd9230a87c7791785f94b2c2f8ebb
Refactor parsing metadata to a common function

Currently the metadata is parsed using
dict(v.split('=', 1) for v in metadata)
and was used in several places, it could
be better if we refactor it to a common
function. It will make it more readable
and easier to use in the future.

Change-Id: Ib0fcaacdbef65cdcb85ecc0c304f4933bef2627c
Closes-Bug: #1503939