Updated openstack/openstack Project: openstack/python-keystoneclient  3ccf42ab3f378215cc803c37c50be6808d258e2e
Iterate over copy of session.adapters keys in Python2/3

Iterate over a copy of session.adapters keys in both Python 2.x and
Python 3.x.  In Python 3.x, keys() is not a copy, and therefore
items can't be popped from it while iterating.

Note that this patch addresses the following error message which
is new in Python 3.5.  That is, you need to be using Python 3.5+
to reproduce this error.

    RuntimeError: OrderedDict mutated during iteration
    https://bugs.python.org/issue24369
    https://hg.python.org/cpython/rev/0d8679858272

Change-Id: Iaa2be0dc8ef26e51ce5e8f50049c9e8f84418ec0
Closes-Bug: #1483872