Updated openstack/openstack Project: openstack-infra/storyboard  15b62368b56cd4b09fc1a9914bd262d056924bf1
Check that the preference being updated is the right one

It should be possible for a key to be a substring in a different
preference key. This patch ensures that the preference used when
updating matches the key exactly rather than getting the first
preference whose key contained the desired key.

Change-Id: I6e31829ceb00a7403bec50b6aa48002e62be1f6e