Updated openstack/openstack Project: openstack/nova-specs  aaeaa454a14f5eada894c690a8cc8ec86c7ff4da
Rework policies for virt-driver CPU thread pinning

https://blueprints.launchpad.net/nova/+spec/virt-driver-cpu-pinning

It was requested that an additional 'prefer' thread policy be added and
the 'avoid' and 'separate' policies be removed.

* The 'prefer' policy corresponds to the existing CPU thread policy.
  Without this change, the policy can be implicitly selected by leaving
  the 'hw:cpu_thread_policy' instance metadata and
  'hw_cpu_thread_policy' image metadata fields blank. This change
  allows the user to explicitly do this, while retaining the existing
  implicit behavior. This is similar to how the 'shared' CPU policy
  works.
* The 'avoid' thread policy provides another way to implement something
  that can already be achieved using availability zones and the policy
  is therefore be dropped.
* The 'separate' thread policy only helps to optimize performance (L2
  cache usage) while the siblings are not occupied by other VMs. This
  temporal optimization provides little value and is therefore dropped.

Some required clarifications on the remaining policies are also
included. Explicitly state whether hyperthreads are required or not
(this was implicit previously, which led to some misinterpretation).

There is a dependency on a bugfix which is required to enable usage of
the 'prefer' policy on hosts with sibling counts that are not dividends
of the overall instance core count. Document this.

Finally, "hyperthreading" is an implementation of SMT. Use the generic
term to illustrate the feature's applicability to other
SMT-implementations, i.e. on POWER.

Change-Id: I260d857921cc2191fbdcef6d8d6fb6af6e886af9