Updated openstack/openstack Project: openstack/requirements  b6193fc059eff6c123e978e968ab1c87818f71a4
Bump PyECLib version to >=1.0.7

This is the first step in updating PyECLib to 1.1.1. Once
this change lands, we plan to update Swift requirements to
>=1.0.7 before updating global-requirements and upper-
constraints to 1.1.1.

Context: The new version of PyECLib introduces native EC
backend named 'liberasurecode_rs_vand' that provides an
alternative to the current 'jerasure_rs_vand' default.
Also brings in a number of bug fixes including those for
issues on big-endian archs.

Change-Id: Ib4bbeb7dfd2ba8f00e269f99f984ade72a77eebc