Updated openstack/openstack Project: openstack-infra/system-config  c19fe6d78f59f9d4e8400ee0b2243971767fec5b
Special case ansible logs

Special case ansible logs which don't have timestamps in console logs as
non multiline events. This should clean up some of our mutltiline event
merging that is happening where it shouldn't.

Note that we make the ansible line not match the multiline filter to
force the previous event to be flushed as is. The ansible line will then
fail the grok filter and become a grok failure as the timestamp is
missing.

Change-Id: I2d3afa3e97dc3c7e02be17268821482c7df60ef3