Updated openstack/openstack Project: openstack-infra/project-config  9d1cba652b1343dbc32c13b18443b499779ef33b
Identity API v3 only jobs non-voting in QA repos

There is an experimental gate job for Identity API
v3 only for DevStack and Tempest repositories,
where the Identity v2 API is disabled.

That job aims to verify that Tempest tests can run
with the described configuration, and consequently
services are compatible with the Identity API v3.

It was made experimental because DevStack was not
using the Identity API v3 to set resources up and
Tempest tests were failing. Now that they are
fixed, the job is passing (see change #213430 as
an example).

This change moves the job from the experimental
pipeline to the main check pipeline for the
DevStack project, making it non-voting for now.

The same change will be done for Tempest after the
Juno EOL date, because there are already too many
tempest jobs running on check.

The goal is, once there is enough confidence on
its stability, to make it voting.

Change-Id: I54f899a9bd2bce94d221933ddaaf35683331766c