Updated openstack/openstack Project: openstack/nova  225715de0095334b70d82ef0556e032e2d1013d1
skip lock_unlock_server test for cells

The cells architecture of doing db replay makes this call
fundamentally racey for cells. The contract is that if you request an
action on a server and it is locked, you get a 409 at the API, and are
done. If you get a 2xx, the resource is not locked, and life is good.

Except in cells, the API db will be updated, and the API will return
all is well. Except the db replay hasn't happened for the cell, so the
additional check in the nova compute for locked blows up.

The fix probably involves the nova-compute check doing a timed retry
under cells because of replication lags.

Change-Id: Ic18407b8372c286d8e258860f3e2bf76eceef2aa
Related-Bug: #1466696