Updated openstack/openstack Project: openstack-infra/project-config  1d3b05b58c4b1b5124cfa2eb8517863fe73cd257
Add ansible-role-nodepool project

In this patch, we are just adding ansible-role-nodepool into the
OpenStack gerrit workflow. After talking with both openstack-infra and
third party CI teams it was decided to deal with which team this would
be part of at a later date.

Additionally, we have not created a launchpad project for this module
or pypi (since we won't release there).

Gerritbot was also not configured at this time, since we don't know
which team we are going to be part of.

Change-Id: I8e9c4008648d55cd6c08ffc87ea54f24c55f9f61
Signed-off-by: Paul Belanger <pabelanger@redhat.com>