Updated openstack/openstack Project: openstack/nova  a6659219f6b5314284b489e6afd20c4720e04d35
clean up regex in tempest-dsvm-cells-rc

The current regex contains a lot of slightly off regex
definitions. Most of them don't hurt, and do still match the correct
way, but they make people believe that things like the * at the end of
matches is important (which it is not).

This cleans this up, plus adds a few comments to help people going
forward.

Change-Id: Ica7fc32fd5a756214db987b95d2305bc74cfb6f1