Updated openstack/openstack Project: openstack/nova  f0882bf2ec3d9ca06dd0afb2eec8846b75ea57be
force_config_drive: StrOpt -> BoolOpt

The 'force_config_drive' option provides a way to force injection of
configuration drives on instances. Previously, this option supported
three values: 'always', 'True' and 'False'. However, the 'always'
value was deprecated in Liberty and can now be removed.

Remove this value and replace the existing 'StrOpt' with a 'BoolOpt'.
This also requires changing some unit tests to reflect the new value
for this parameter.

UpgradeImpact: This change removes support for the 'always' option.
    Configuration files providing this value will need to be updated
    accordingly.
DocImpact: This option is now a BoolOpt and documentation will need to
    be updated accordingly.
Change-Id: I0b4030ec1336fb1f2e721570c99a27a979e2847c