Updated openstack/openstack Project: openstack/nova  c52f18bcf1ab6820ebb410cf73ea69ee3568c26b
Properly log BlockDeviceMappingList in _create_block_device_mapping

The block_device_mapping list parameter is a BlockDeviceMappingList
object, and list objects don't repr the sub-objects so it's useless
information in the logs. Cast to a list to get the sub-objects repr'ed
in the logs.

The bdm list of dicts from the API request is converted to a
BlockDeviceMappingList in _check_and_transform_bdm which is called from
_create_instance before _create_block_device_mapping.

Closes-Bug: #1513934

Change-Id: Id709c0bc9d0ad980d0b36973f3a52034003170af