Updated openstack/openstack Project: openstack/heat-templates  be3604a37427dc556c7c714bedc4bcc9ab11882b
Move deployed dir from /var/run to /var/lib

Deployments should not be run again when a server is rebooted because
heat will likely not be listening for the deployment signals. This
change moves the deployed directory from /var/run/heat-config/deployed
to /var/lib/heat-config/deployed so that deployed state is persisted
across reboots.

There is migration logic to move the existing files from the old to the
new location, this will only be run on the first run of 55-heat-config
after its package is updated.

Change-Id: I3d305a4ac5b68c29037760682d37e5b9a530828e
Closes-Bug: #1513220