Updated openstack/openstack Project: openstack/keystone  82afe4be209b94758d26c78cf17e3b989f842c8c
Pass kwargs when using revoke_api.list_events()

The list_events method takes a keyword argument for last_fetch. One of the
tests in test_revoke.py was passing it as positional. I noticed this when
refactoring sql.DateTime formats out of the revocation events schema. This it
just a minor commit to fix it.

Change-Id: Ic62f2f2d928f7f09d98f12ce74bb2db7eae8f673